customer-package/lambda/requirements.txt
PyJWT==2.8.0
cryptography==41.0.7
boto3>=1.26.0
